split out the fuzz tests (#202)
