+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod rewriting;
 pub mod egraphs;
 pub mod util;
 pub mod formats;
+pub mod test_utils;
 
 pub use {
     compression::*,
@@ -12,6 +13,7 @@ pub use {
     util::*,
     formats::*,
     lambdas::*,
+    test_utils::*,
 };
 
 pub use colorful::{Color,Colorful,RGB};
